% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cohend}
\alias{cohend}
\title{Compute Cohen's d}
\usage{
cohend(diff, ns, sds, conf = 0.95)
}
\arguments{
\item{diff}{Numeric difference value}

\item{ns}{Numeric vector containing the sample sizes of the two groups}

\item{sds}{Numeris vector containing the sample standard deviations of the
two groups (at the point of interest if taken on functional data).}

\item{conf}{Numeric value giving the desired confidence level.}
}
\value{
A vector of length 3 containing the lower bound, the effect size,
and the upper bound in that order.
}
\description{
Calculates Cohen's d, a measure of effect size as well as confidence interval
bounds.
}
\details{
Arguments are given in vector form so as to be more amenable to the
structure of the package output. Confidence interval calculation comes from
Hedge and Olkin (2014).
}
