% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cov_fda}
\alias{cov_fda}
\title{Covariate Incorporated Bayesian FDA}
\usage{
cov_fda(data_mat, cov_mat, p = 20, niter = 10000, nburn = 1000)
}
\arguments{
\item{data_mat}{matrix of observed curves organized into columns}

\item{cov_mat}{matrix of observed covariates with covariates in rows}
}
\value{
List of MCMC chains. Primary interest may lie in analyzing rows of B_chains.
}
\description{
Fits a special case of a functional regression model
}
