% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{extract_sig_area}
\alias{extract_sig_area}
\title{Extract areas of significance}
\usage{
extract_sig_area(trip)
}
\arguments{
\item{trip}{Matrix with three columns corresponding to the lower bound (
first column), average curve (second column), and upper bound (third column).}
}
\value{
4xk Matrix where each row is a significance region and the columns
correspond to the beginning index, ending index, maximum location, and max value.
}
\description{
Given a 3-column matrix corresponding to the lower, average, and upper curves,
extract the regions of significance, as well as the locations and value of
the largest extrema.
}
