% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{multi_group_fda}
\alias{multi_group_fda}
\title{Multi-Group Bayesian FDA}
\usage{
multi_group_fda(data_list, p = 20, niter = 10000, nburn = 1000)
}
\arguments{
\item{data_list}{list of group matrices where columns correspond to observed curves}
}
\value{
List of MCMC chains. Primary interest may lie in analyzing beta_grp chains.
}
\description{
Fits a multi-group Bayesian FDA using penalized P-splines
}
\details{
Fits Bayesian P-splines model. The model can be given by:
\deqn{y_{ij} \sim N(H\beta_{ij}, \sigma^2)\\
\beta_{ij} \sim N(\beta_j, \tau^2P^{-1})\\
p(\beta_j) \propto 1; p(\sigma^2) \propto \frac{1}{\sigma^2}; p(\tau^2) \propto \frac{1}{\tau^2}}
}
